requirements.txt
black==23.9.0
click==8.1.7
pytest==7.4.2
pytest-cov==4.1.0
pylint==2.17.5
nbval==0.10.0


farm-haystack[inference]==1.21.0
streamlit==1.27.1
torch